reorganize `bee-packable`
