+++ bee-common/bee-packable/src/error.rs
@@ -87,43 +87,6 @@ impl<T> From<Infallible> for UnknownTagError<T> {
     }
 }
 
-/// Semantic error raised while unpacking a dynamically-sized sequences that use a type different than `usize` for their
-/// length-prefix.
-#[derive(Debug)]
-pub enum UnpackPrefixError<T, E> {
-    /// Semantic error raised while unpacking an element of the sequence.
-    /// Typically this is [`Packable::UnpackError`](crate::Packable).
-    Packable(T),
-    /// Semantic error raised when the length prefix cannot be unpacked.
-    Prefix(E),
-}
-
-impl<E> UnpackPrefixError<Infallible, E> {
-    /// Projects the value to the [`Prefix`](UnpackPrefixError::Prefix) variant.
-    pub fn into_prefix(self) -> E {
-        match self {
-            Self::Packable(err) => match err {},
-            Self::Prefix(err) => err,
-        }
-    }
-}
-
-impl<T, E> UnpackPrefixError<T, E> {
-    /// Returns the contained [`Packable`](UnpackPrefixError::Packable) value or computes it from a closure.
-    pub fn unwrap_packable_or_else<V: Into<T>>(self, f: impl FnOnce(E) -> V) -> T {
-        match self {
-            Self::Packable(err) => err,
-            Self::Prefix(err) => f(err).into(),
-        }
-    }
-}
-
-impl<T, E> From<T> for UnpackPrefixError<T, E> {
-    fn from(err: T) -> Self {
-        Self::Packable(err)
-    }
-}
-
 /// Error type to be raised when [`SliceUnpacker`](`crate::unpacker::SliceUnpacker`) does not have enough bytes to
 /// unpack something or when [`SlicePacker`]('crate::packer::SlicePacker') does not have enough space to pack something.
 #[derive(Debug)]
