+++ bee-common/bee-packable/src/packable/option.rs
@@ -1,7 +1,13 @@
 // Copyright 2021 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
-use crate::{error::UnpackError, packer::Packer, unpacker::Unpacker, Packable, UnpackErrorExt};
+//! Types and utilities related to packing and unpacking [`Option`] values.
+use crate::{
+    error::{UnpackError, UnpackErrorExt},
+    packer::Packer,
+    unpacker::Unpacker,
+    Packable,
+};
 
 /// Error type raised when a semantic error occurs while unpacking an option.
 #[derive(Debug)]
